+++ CHANGES.md
@@ -15,3 +15,4 @@
 ### Bug Fixes
 
 * Backslashes in front of quotation mark were escaping the quotation mark, even if the backslash itself was escaped.
+* Unescaped double-quote marks weren't properly handled, e.g., (parser "A = '\"'").